// app/demos/webgl_multiple_elements/type.ts
export interface MultipleElementType {
    /**
     * dom-id
     */
    id: string;
    /**
     * title：标题
     */
    title: string;
    /**
     * desc：描述
     */
    desc: string;
    /**
     * 几何模型
     */
    geometry: any;
    /**
     * 材质
     */
    material: any;
}